Fix migrations for data column

+++ src/migrations/2023_07_06_000000_add_data_to_medias_table.php
@@ -13,9 +13,11 @@ class AddDataToMediasTable extends Migration
      */
     public function up()
     {
-        Schema::table(config('mediatheque.table_prefix').'medias', function (Blueprint $table) {
-            $table->json('data')->nullable()->after('name');
-        });
+        if (!Schema::hasColumn('mediatheque_medias', 'data')) {
+            Schema::table(config('mediatheque.table_prefix') . 'medias', function (Blueprint $table) {
+                $table->json('data')->nullable()->after('name');
+            });
+        }
     }
 
     /**
@@ -25,8 +27,10 @@ class AddDataToMediasTable extends Migration
      */
     public function down()
     {
-        Schema::table(config('mediatheque.table_prefix').'medias', function (Blueprint $table) {
-            $table->dropColumn('data');
-        });
+        if (Schema::hasColumn('mediatheque_medias', 'data')) {
+            Schema::table(config('mediatheque.table_prefix') . 'medias', function (Blueprint $table) {
+                $table->dropColumn('data');
+            });
+        }
     }
 }
